# packaging-catalog-bc/copilot/GPI-PackagingQuoteAction.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GPI Packaging Quote Copilot Actions",
    "version": "1.1.0",
    "description": "Controlled Copilot-facing contract for preparing and executing Business Central packaging quote actions. Business Central remains authoritative for pricing, guardrails, approvals, quote status, and action eligibility. Approval and rejection may include a confirmation-bound decision note."
  },
  "tags": [
    {
      "name": "Packaging Quote Actions",
      "description": "Prepare and execute controlled Business Central packaging quote actions."
    }
  ],
  "paths": {
    "/packagingQuoteActions/prepare": {
      "post": {
        "operationId": "PreparePackagingQuoteAction",
        "summary": "Prepare a packaging quote action",
        "description": "Reads current Business Central state and prepares the requested action. For approve or reject, an optional decisionNote may be proposed and bound into the confirmation token without being written during Prepare. This operation does not change Business Central or Spiro.",
        "tags": [
          "Packaging Quote Actions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrepareActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Current action preparation result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrepareActionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/packagingQuoteActions/execute": {
      "post": {
        "operationId": "ExecutePackagingQuoteAction",
        "summary": "Execute a confirmed packaging quote action",
        "description": "Revalidates live Business Central state and the confirmation token. When a confirmation-bound decisionNote is supplied for approve or reject, the controlled flow writes that note first, revalidates action eligibility, and then delegates the Business Central bound action.",
        "tags": [
          "Packaging Quote Actions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Confirmed action executed and resulting Business Central state returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteActionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PackagingQuoteAction": {
        "type": "string",
        "enum": [
          "evaluate",
          "readyForReview",
          "approve",
          "reject",
          "reopen"
        ],
        "description": "Exact Business Central bound action name."
      },
      "PrepareActionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "quoteNo",
          "action"
        ],
        "properties": {
          "quoteNo": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "description": "Business Central packaging quote entry number."
          },
          "action": {
            "$ref": "#/components/schemas/PackagingQuoteAction"
          },
          "decisionNote": {
            "type": "string",
            "minLength": 1,
            "nullable": true,
            "description": "Optional decision note proposed for approve or reject. Prepare remains read-only; when required, the exact note is bound into the confirmation token."
          }
        },
        "example": {
          "quoteNo": 55,
          "action": "approve",
          "decisionNote": "Approved after review of fixed price exception."
        }
      },
      "ExecuteActionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "quoteNo",
          "action",
          "confirmationToken"
        ],
        "properties": {
          "quoteNo": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "action": {
            "$ref": "#/components/schemas/PackagingQuoteAction"
          },
          "confirmationToken": {
            "type": "string",
            "minLength": 1,
            "description": "Exact state-bound token returned by the immediately preceding Prepare operation."
          },
          "decisionNote": {
            "type": "string",
            "minLength": 1,
            "nullable": true,
            "description": "For approve or reject when a decision note was supplied during Prepare, send the exact same note with the confirmation token. The note is part of the token-bound confirmation state."
          }
        },
        "example": {
          "quoteNo": 55,
          "action": "approve",
          "confirmationToken": "GPI-Q55-APPROVE-example",
          "decisionNote": "Approved after review of fixed price exception."
        }
      },
      "PrepareActionResponse": {
        "type": "object",
        "required": [
          "schemaVersion",
          "mode",
          "quoteNo",
          "currentStatus",
          "requestedAction",
          "isAllowed",
          "confirmationRequired",
          "allowedWriteActions",
          "contractVersion",
          "writeExecution"
        ],
        "properties": {
          "schemaVersion": {
            "type": "string",
            "example": "1.0"
          },
          "mode": {
            "type": "string",
            "enum": [
              "Prepare"
            ]
          },
          "quoteNo": {
            "type": "integer",
            "format": "int32"
          },
          "quoteId": {
            "type": "string",
            "format": "uuid"
          },
          "customerNo": {
            "type": "string"
          },
          "customerName": {
            "type": "string"
          },
          "currentStatus": {
            "type": "string"
          },
          "requestedAction": {
            "$ref": "#/components/schemas/PackagingQuoteAction"
          },
          "isAllowed": {
            "type": "boolean"
          },
          "confirmationRequired": {
            "type": "boolean"
          },
          "confirmationText": {
            "type": "string",
            "nullable": true
          },
          "expectedResultStatus": {
            "type": "string",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "confirmationToken": {
            "type": "string",
            "nullable": true,
            "description": "Returned only when the action is permitted and confirmation is required."
          },
          "allowedWriteActions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PackagingQuoteAction"
            }
          },
          "contractVersion": {
            "type": "string"
          },
          "recommendedNextAction": {
            "type": "string"
          },
          "hasPricingExceptions": {
            "type": "boolean"
          },
          "primaryPricingExceptionStatus": {
            "type": "string"
          },
          "primaryPricingExceptionMessage": {
            "type": "string"
          },
          "primaryApprover": {
            "type": "string"
          },
          "writeExecution": {
            "$ref": "#/components/schemas/WriteExecutionPreparation"
          },
          "decisionNoteRequired": {
            "type": "boolean",
            "description": "True when the requested approve or reject action requires a decision note under the current Business Central rules."
          },
          "decisionNoteProvided": {
            "type": "boolean",
            "description": "True when the caller supplied a nonblank decisionNote with Prepare."
          },
          "decisionNote": {
            "type": "string",
            "nullable": true,
            "description": "Normalized proposed decision note. Prepare does not persist this value."
          },
          "decisionNoteWillBeWritten": {
            "type": "boolean",
            "description": "True when confirmed Execute will persist the proposed decision note before invoking approve or reject."
          },
          "decisionNoteEnablesAction": {
            "type": "boolean",
            "description": "True when the supplied decision note satisfies the missing prerequisite that otherwise prevents approve or reject."
          }
        }
      },
      "WriteExecutionPreparation": {
        "type": "object",
        "required": [
          "permitted",
          "executeMode",
          "confirmationTokenRequired"
        ],
        "properties": {
          "permitted": {
            "type": "boolean"
          },
          "executeMode": {
            "type": "string",
            "enum": [
              "Execute"
            ]
          },
          "confirmationTokenRequired": {
            "type": "boolean"
          }
        }
      },
      "ExecuteActionResponse": {
        "type": "object",
        "required": [
          "schemaVersion",
          "mode",
          "quoteNo",
          "requestedAction",
          "executed",
          "statusBefore",
          "statusAfter",
          "resultingAllowedWriteActions",
          "contractVersion"
        ],
        "properties": {
          "schemaVersion": {
            "type": "string",
            "example": "1.0"
          },
          "mode": {
            "type": "string",
            "enum": [
              "Execute"
            ]
          },
          "quoteNo": {
            "type": "integer",
            "format": "int32"
          },
          "quoteId": {
            "type": "string",
            "format": "uuid"
          },
          "requestedAction": {
            "$ref": "#/components/schemas/PackagingQuoteAction"
          },
          "executed": {
            "type": "boolean"
          },
          "statusBefore": {
            "type": "string"
          },
          "statusAfter": {
            "type": "string"
          },
          "expectedResultStatus": {
            "type": "string",
            "nullable": true
          },
          "resultingAllowedWriteActions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PackagingQuoteAction"
            }
          },
          "recommendedNextAction": {
            "type": "string"
          },
          "contractVersion": {
            "type": "string"
          },
          "decisionNoteWritten": {
            "type": "boolean",
            "description": "True when Execute persisted the confirmation-bound decision note before the Business Central action."
          },
          "decisionNote": {
            "type": "string",
            "nullable": true,
            "description": "Decision note persisted by this Execute request, when applicable."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Stable machine-readable error category."
          },
          "message": {
            "type": "string",
            "description": "Human-readable refusal or failure reason."
          },
          "quoteNo": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "action": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PackagingQuoteAction"
              }
            ],
            "nullable": true
          },
          "retryPrepare": {
            "type": "boolean",
            "description": "True when the caller should prepare again before attempting execution."
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Request failed validation or violates a hard safety boundary.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Conflict": {
        "description": "Current Business Central state no longer permits the requested operation or the confirmation token is stale.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ServerError": {
        "description": "Unexpected integration or downstream Business Central failure.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "x-gpi-implementation": {
    "milestone": "0.43",
    "environment": "Sandbox_NoZetadocs_UAT",
    "businessCentralAuthority": true,
    "prepareComponent": "Prepare-GPIPackagingQuoteActionUAT.ps1",
    "executeComponent": "Invoke-GPIPackagingQuoteActionUAT.ps1",
    "flowComponent": "Invoke-GPIPackagingQuoteCopilotFlowUAT.ps1",
    "protectedQuoteNumbers": [
      67
    ],
    "confirmationModel": "Prepare returns a state-bound confirmation token. When decisionNote is supplied, the normalized note is also bound into that token. Execute must receive the exact token and same decisionNote and must revalidate live Business Central state before writing.",
    "hostingStatus": "Local UAT FastAPI HTTP adapter implemented. Execute is disabled by default and requires GPI_ENABLE_COPILOT_EXECUTE=1.",
    "decisionNoteComponent": "Set-GPIPackagingQuoteDecisionNoteUAT.ps1",
    "decisionNoteModel": "Prepare is read-only. For approve/reject, an optional decisionNote may satisfy the missing action prerequisite. Confirmed Execute writes the token-bound note only while the quote is Ready, verifies persistence, re-reads Business Central action eligibility, then delegates the bound action."
  }
}